requirements.txt
streamlit
streamlit-qrcode-scanner
pandas
pdfplumber
requests
Pillow
openpyxl
selenium
webdriver-manager
opencv-python-headless
numpy
pyzbar

